+++ packages/app-desktop/src/features/editor/text-editor.tsx
@@ -49,6 +49,7 @@ export const TextEditor = ({ initialValue, onChange }: TextEditorProps) => {
                 //if (event.dataTransfer?.types.includes("note_id"))
                 //console.log("Linking to note");
               },
+              handlePaste(view, event, slice) {},
             }}
             onUpdate={({ editor }) => {
               onChange?.call(null, editor.getJSON());